refactor : 무분별한 디렉토리가 가독성을 떨어뜨린다고 판단, 디렉토리 정리
