+++ src/main/java/edu/nextstep/lottoauto/domain/ticketmaker/CustomNumbersMaker.java
@@ -1,4 +1,4 @@
-package edu.nextstep.lottoauto.ticketmaker;
+package edu.nextstep.lottoauto.domain.ticketmaker;
 
 import java.util.Arrays;
 import java.util.List;
