+++ src/test/java/edu/nextstep/lottoauto/domain/AutoNumbersMakerTest.java
@@ -1,6 +1,8 @@
-package edu.nextstep.lottoauto.ticketmaker;
+package edu.nextstep.lottoauto.domain;
 
 
+import edu.nextstep.lottoauto.domain.ticketmaker.AutoNumbersMaker;
+import edu.nextstep.lottoauto.domain.ticketmaker.NumbersMaker;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
